+++ tests/core/unit_test.py
@@ -10,8 +10,9 @@ def test_unit_e_singleton() -> None:
     assert Unit() is Unit()
 
 
-def test_unit_str() -> None:
+def test_unit_str_e_repr() -> None:
     assert str(UNIT) == 'Unit - void'
+    assert repr(UNIT) == 'UNIT'
 
 
 def test_unit_sobrevive_ao_pickle() -> None: